+++ src/app/app.component.ts
@@ -3,8 +3,7 @@ import { from, fromEvent } from "rxjs"
 
 @Component({
   selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss']
+  template: '<router-outlet></router-outlet>'
 })
 export class AppComponent implements OnInit {
