+++ src/app/pages/home/home.module.ts
@@ -1,13 +1,20 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { HomeRoutingModule } from './home-routing.module';
-import { HomeComponent } from './home/home.component';
-import { ProductService } from '../../services/product.service';
+import { HomeComponent } from './home.component';
+import { ProductService } from '../../shared/services/product.service';
 
 import { MatCardModule } from "@angular/material/card";
 import { MatButtonModule } from "@angular/material/button";
 import { HttpClientModule } from '@angular/common/http';
+import { Routes, RouterModule } from '@angular/router';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: HomeComponent
+  }
+];
 
 @NgModule({
   declarations: [HomeComponent],
@@ -16,7 +23,8 @@ import { HttpClientModule } from '@angular/common/http';
     HomeRoutingModule,
     MatCardModule,
     MatButtonModule,
-    HttpClientModule
+    HttpClientModule,
+    RouterModule.forChild(routes)
   ],
   providers: [
     ProductService
