--- src/app/pages/list/list-routing.module.ts
@@ -1,16 +1,0 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-import { ListComponent } from './list/list.component';
-
-const routes: Routes = [
-  {
-    path: '',
-    component: ListComponent
-  }
-];
-
-@NgModule({
-  imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
-})
-export class ListRoutingModule { }
