+++ src/app/pages/list/list.component.ts
