+++ src/app/pages/list/list.module.ts
@@ -1,15 +1,21 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { ListRoutingModule } from './list-routing.module';
-import { ListComponent } from './list/list.component';
+import { ListComponent } from './list.component';
+import { Routes, RouterModule } from '@angular/router';
 
+const routes: Routes = [
+  {
+    path: '',
+    component: ListComponent
+  }
+];
 
 @NgModule({
   declarations: [ListComponent],
   imports: [
     CommonModule,
-    ListRoutingModule
+    RouterModule.forChild(routes)
   ]
 })
 export class ListModule { }
