+++ src/app/pages/pages.component.ts
@@ -1,9 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: 'app-pages',
-  templateUrl: './pages.component.html',
-  styleUrls: ['./pages.component.scss']
+  templateUrl: './pages.component.html'
 })
 export class PagesComponent implements OnInit {
