+++ src/app/pages/pages.module.ts
@@ -1,15 +1,27 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { PagesRoutingModule } from './pages-routing.module';
-import { PagesComponent } from './pages/pages.component';
+import { PagesComponent } from './pages.component';
+import { Routes, RouterModule } from '@angular/router';
 
+const routes: Routes = [
+  {
+    path: '',
+    component: PagesComponent,
+    children: [
+      {path: '', redirectTo: 'home', pathMatch: 'full'},
+      {path: 'home', loadChildren: () => import('./home/home.module').then(m => m.HomeModule)},
+      {path: 'admin', loadChildren: () => import('./admin/admin.module').then(m => m.AdminModule)},
+      {path: 'list', loadChildren: () => import('./list/list.module').then(m => m.ListModule)}
+    ]
+  }
+];
 
 @NgModule({
   declarations: [PagesComponent],
   imports: [
     CommonModule,
-    PagesRoutingModule
+    RouterModule.forChild(routes)
   ]
 })
 export class PagesModule { }
