+++ src/app/shared/services/product.service.ts
@@ -1,25 +1,28 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
+import { environment } from '../../environments/environment'
 
 @Injectable()
 export class ProductService {
 
+  url = environment.app.apiBaseUrl
+
   constructor(private http: HttpClient) { }
 
   public getProducts(): Observable<any> {
-    return this.http.get('https://angular-course-ff5dd.firebaseio.com/products.json');
+    return this.http.get(`${this.url}/products.json`);
   }
 
   public addProduct(product: any): Observable<any> {
-    return this.http.post('https://angular-course-ff5dd.firebaseio.com/products.json', product);
+    return this.http.post(`${this.url}/products.json`, product);
   }
 
   public deleteProduct(id: any): Observable<any> {
-    return this.http.delete(`https://angular-course-ff5dd.firebaseio.com/products/${id}.json`);
+    return this.http.delete(`${this.url}/products/${id}.json`);
   }
 
   public updateProduct(id: any, product: any): Observable<any> {
-    return this.http.put(`https://angular-course-ff5dd.firebaseio.com/products/${id}.json`, product);
+    return this.http.put(`${this.url}/products/${id}.json`, product);
   }
 }
