+++ src/environments/environment.prod.ts
@@ -1,3 +1,10 @@
 export const environment = {
-  production: true
+  production: true,
+  app: {
+    apiBaseUrl: 'https://angular-course-ff5dd.firebaseio.com'
+  },
+  auth: {
+    apiBaseUrl: 'https://identitytoolkit.googleapis.com',
+    key: 'AIzaSyDDCNbHMWv5kssnDbq4kWPCb550rNUi7HU'
+  }
 };
