+++ src/environments/environment.ts
@@ -3,7 +3,14 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  app: {
+    apiBaseUrl: 'https://angular-course-ff5dd.firebaseio.com'
+  },
+  auth: {
+    apiBaseUrl: 'https://identitytoolkit.googleapis.com',
+    key: 'AIzaSyDDCNbHMWv5kssnDbq4kWPCb550rNUi7HU'
+  }
 };
 
 /*
